refactor: move src to project root
