--- api/.dockerignore
@@ -1,1 +1,0 @@
-src/first-run-false
