--- api/pyproject.toml
@@ -1,51 +1,0 @@
-[tool.poetry]
-name = "api"
-version = "0.1.0"
-description = "API running jobs in development framework"
-authors = [""]
-license = "MIT"
-
-[tool.poetry.dependencies]
-python = "^3.10.4"
-dmss-api = "^0.3.17"
-redis = "^4.3.4"
-APScheduler = "^3.9.1"
-docker = "^5.0.3"
-fastapi = "^0.84.0"
-gunicorn = "^20.1.0"
-uvicorn = {extras = ["standard"], version = "^0.18.2"}
-
-
-
-[tool.poetry.dev-dependencies]
-pytest = "^7.1.1"
-deepdiff = "^5.8.1"
-pre-commit = "^2.19.0"
-pygments = "^2.12.0"
-coverage = "^6.4.2"
-flake8 = "^5.0.4"
-bandit = "^1.7.4"
-safety = "^2.1.1"
-black = "^22.6.0"
-
-[tool.isort]
-profile = "black"
-
-[tool.black]
-line-length = 119
-target-version = ['py10']
-exclude = '''
-/(
-    \.eggs
-  | \.git
-  | \.hg
-  | \.mypy_cache
-  | \.tox
-  | \.venv
-  | \dmss
-)/
-'''
-
-[build-system]
-requires = ["poetry-core>=1.0.0"]
-build-backend = "poetry.core.masonry.api"
