--- api/src/enums.py
@@ -1,14 +1,0 @@
-from enum import Enum
-
-
-class SIMOS(Enum):
-    BLUEPRINT = "system/SIMOS/Blueprint"
-    BLOB = "system/SIMOS/Blob"
-    BLUEPRINT_ATTRIBUTE = "system/SIMOS/BlueprintAttribute"
-    APPLICATION = "system/SIMOS/Application"
-    ATTRIBUTE_TYPES = "system/SIMOS/AttributeTypes"
-    UI_RECIPE = "system/SIMOS/UiRecipe"
-    STORAGE_RECIPE = "system/SIMOS/StorageRecipe"
-    PACKAGE = "system/SIMOS/Package"
-    ENTITY = "system/SIMOS/Entity"
-    DATASOURCE = "datasource"
