+++ docker-compose.override.yml
@@ -1,12 +1,12 @@
 version: "3.4"
 
 services:
-  api:
+  job-api:
     build:
       target: development
-    image: api-dev
+    image: job-api-dev
     volumes:
-      - ./api/src/:/code/src
+      - ./src/:/code/src
       - /var/run/docker.sock:/var/run/docker.sock # Needed for docker-in-docker jobs
       # - ../data-modelling-storage-service/gen/dmss_api:/dmss_api
       # - ../data-modelling-storage-service/src/home:/code/src/dmss/src/home
