+++ src/default_job_handlers/omnia_classic_azure_container_instances/OmniaClassicContainerInstance.json
